backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
psycopg2-binary==2.9.10
alembic==1.14.1
pydantic[email]==2.10.4
python-multipart==0.0.19
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
PyJWT==2.10.1
boto3==1.35.86
qrcode[pil]==8.0
reportlab==4.2.5
Pillow==11.1.0
deepface==0.0.93
tf-keras==2.18.0
httpx==0.28.1
python-dotenv==1.0.1
aiofiles==24.1.0
pytest==8.3.4
pytest-asyncio==0.24.0
